+++ Sources/NodeUIEngine/NUIE_SvgDrawingContext.hpp
@@ -41,13 +41,13 @@ private:
 class SvgDrawingContext : public DrawingContext
 {
 public:
-	SvgDrawingContext (double width, double height);
+	SvgDrawingContext (int width, int height);
 
 	std::wstring		GetAsString () const;
 
 	virtual void		Resize (int newWidth, int newHeight) override;
-	virtual double		GetWidth () const override;
-	virtual double		GetHeight () const override;
+	virtual int			GetWidth () const override;
+	virtual int			GetHeight () const override;
 	virtual void		BeginDraw () override;
 	virtual void		EndDraw () override;
 	virtual bool		NeedToDraw (ItemPreviewMode mode) override;
@@ -64,8 +64,8 @@ public:
 
 private:
 	SvgBuilder	svgBuilder;
-	double		width;
-	double		height;
+	int			width;
+	int			height;
 };
 
 std::wstring ReplaceAll (const std::wstring& string, const std::wstring& from, const std::wstring& to);
